Change to new cluster selection
