--- pkg/scheduler/core/spreadconstraints/select_clusters.go
@@ -1,38 +1,0 @@
-/*
-Copyright 2022 The Karmada Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package spreadconstraint
-
-import (
-	"fmt"
-	clusterv1alpha1 "github.com/karmada-io/karmada/pkg/apis/cluster/v1alpha1"
-)
-
-// SelectBestClusters selects the best clusters based on the provided selection context and factory name.
-func SelectBestClusters(ctx SelectionCtx, name string) ([]*clusterv1alpha1.Cluster, error) {
-	if name == "" {
-		name = DefaultSelectionFactoryName
-	}
-	factory, ok := SelectionRegistry[name]
-	if !ok {
-		return nil, fmt.Errorf("selection %s is not registered", name)
-	}
-	selection, err := factory.Create(ctx)
-	if err != nil {
-		return nil, err
-	}
-	return selection.Elect()
-}
